ml-worker-python/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
httpx==0.27.0
python-dotenv==1.0.1

# ML / Data Science
numpy>=2,<3
scipy>=1.13,<2
scikit-learn>=1.5,<2
pandas>=2.2,<3
joblib>=1.3,<2

# Explainability (Python 3.12 wheels require numpy>=2)
shap>=0.50.0

# NLP / Sentiment
# (Using lexicon-based proxy to save storage space)

# Google Gemini
google-genai==0.5.0

# Observability
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0

# Security
python-jose[cryptography]==3.3.0
cryptography==42.0.5

# Utilities
requests==2.31.0
python-multipart==0.0.9
aiofiles==23.2.1
feedparser>=6.0.11          # Google News RSS parsing (fallback to stdlib xml if unavailable)

# Document parsing (Pillar 1)
PyMuPDF==1.24.7
pdfplumber==0.11.4

# OCR (optional; will degrade gracefully if tesseract binary missing)
pytesseract==0.3.10
Pillow==10.4.0

# CAM export
fpdf2==2.7.9
python-docx==1.1.2

# PDF Generation
reportlab==4.4.10
matplotlib==3.10.8


# OCR-LLM Dependencies
torch
transformers>=4.45.0
pdf2image
accelerate
torchvision
qwen-vl-utils
